feat(opatio): added ability to store vectors in each cell
each cell in an opatio file can now store either a scalar or a vector. This is defined by the size field in the table index for that card. They are stored as a contiguous block of memory of float64s * size (i.e if size is 1 then this is exactly equivilent to storing scalars). The python module has been updated to handle this.