build(qhull): added patch from https://github.com/microsoft/vcpkg/pull/37765
qhull version 8.0.2 uses template parameters from its class templates in its constructor. This was allowed in versions of c++ prior to C++20 but we target C++23 and this results in compile time errors when building with gcc versions >= 11. This patch was developed by microsoft to allow the qhull c++ api to compile against newer c++ standards.